Check for installation
There seem to be a problem here. The plugin is disabled during maintenance but not during installation and it's apparently causing the installation to crash.

I've patched it with a very quick and dirt low level check, but I reckon the flaw is likely to be in the plugin design.